+++ src/main/java/calculator/LocalProduceCalculator.java
@@ -7,8 +7,7 @@ public class LocalProduceCalculator {
      * @param weight Amount of gram food bought.
      * @return (Amount of grams bought imported food) - (Amount of grams bought local produced food) = amount of grams CO2 saved
      */
-
-    public static int produce(int weight) {
+    public static double produce(int weight) {
         //Calculate amount of kg CO2 saved.
         Double localFood = 232.0 * weight;
         Double importFood = 560.0 * weight;
@@ -17,7 +16,7 @@ public class LocalProduceCalculator {
 
         doubleResult = doubleResult /1000000;
         doubleResult = Math.round(doubleResult*100.00) / 100.00;
-        return doubleResult.intValue();
+        return doubleResult;
     }
