+++ src/test/java/calculator/BusCalculatorTest.java
@@ -31,7 +31,7 @@ public class BusCalculatorTest {
      */
     @Test
     public void TestBus() {
-        int result = BusCalculator.bus(30);
+        Double result = BusCalculator.bus(30);
         Assert.assertEquals(3.8, result, 1);
     }
 
@@ -42,7 +42,7 @@ public class BusCalculatorTest {
      */
     @Test
     public void TestBusCar() throws MalformedURLException {
-        int result = BusCalculator.carBus(20, 3.1);
+        Double result = BusCalculator.carBus(20, 3.1);
         Assert.assertEquals(2.5, result, 1);
     }
